refactor: extract VendorsPage logic into feature-local hooks
Extract 3 custom hooks from VendorsPage coordinator to reduce it
from 1018 to 536 lines:

- useContactChildrenData: contact email/phone fetching, lookup maps,
  contactColumns definition
- useVendorEditCollections: edit sub-collections (emails, phones,
  addresses, taxIds), sync effects, dirty-check, save handler,
  inline row renderers
- useVendorContactDialogs: contact CRUD dialogs, forms, password
  popover, archive/restore, selection, filtering

Also consolidates vendor-level dialog state using useDialogState
(viewDialog, createDialog, editDialog, importDialog, resetPwDialog).
